+++ packages/actor-memory-expression/tsup.config.ts
@@ -1,0 +1,4 @@
+import { createTsupConfig } from '../../scripts/tsup.config';
+
+// eslint-disable-next-line import/no-default-export
+export default createTsupConfig({});