+++ packages/json_schemas/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.1](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.12.0...@apify/json_schemas@0.12.1) (2026-01-27)
+
+**Note:** Version bump only for package @apify/json_schemas
+
+
+
+
+
+# [0.12.0](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.11.1...@apify/json_schemas@0.12.0) (2026-01-26)
+
+
+### Features
+
+* multiple datasets schema ([#588](https://github.com/apify/apify-shared-js/issues/588)) ([4cfbc6f](https://github.com/apify/apify-shared-js/commit/4cfbc6fc5d1d70df3900d249005fb046fb8953bf))
+
+
+
+
+
+## [0.11.1](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.11.0...@apify/json_schemas@0.11.1) (2026-01-20)
+
+**Note:** Version bump only for package @apify/json_schemas
+
+
+
+
+
+# [0.11.0](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.10.0...@apify/json_schemas@0.11.0) (2026-01-11)
+
+
+### Features
+
+* **json_schemas:** Remove storages.requestQueue from actor definition ([#586](https://github.com/apify/apify-shared-js/issues/586)) ([34816f4](https://github.com/apify/apify-shared-js/commit/34816f4f6c2bb8395803c5943b394b01f32bc9c1))
+
+
+
+
+
+# [0.10.0](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.9.2...@apify/json_schemas@0.10.0) (2025-12-18)
+
+
+### Features
+
+* **json_schemas:** Add `dockerContextDir` and `changelog` to actor schema ([#584](https://github.com/apify/apify-shared-js/issues/584)) ([0c40567](https://github.com/apify/apify-shared-js/commit/0c4056752bc08eeb371190991186ca730511bd3f))
+
+
+
+
+
+## [0.9.2](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.9.1...@apify/json_schemas@0.9.2) (2025-11-27)
+
+**Note:** Version bump only for package @apify/json_schemas
+
+
+
+
+
+## [0.9.1](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.9.0...@apify/json_schemas@0.9.1) (2025-11-27)
+
+**Note:** Version bump only for package @apify/json_schemas
+
+
+
+
+
+# [0.9.0](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.8.0...@apify/json_schemas@0.9.0) (2025-11-27)
+
+
+### Features
+
+* **input_schema:** Custom error messages ([#567](https://github.com/apify/apify-shared-js/issues/567)) ([cded259](https://github.com/apify/apify-shared-js/commit/cded2591d13de1760807fdc6c0103af0a03b5bc2))
+
+
+
+
+
 # [0.8.0](https://github.com/apify/apify-shared-js/compare/@apify/json_schemas@0.7.0...@apify/json_schemas@0.8.0) (2025-11-18)